Removed non-downloading commands from download_raw_foil_data
