+++ analysis/neutron/download_raw_foil_data.py
@@ -2,19 +2,6 @@ from pathlib import Path
 import zipfile
 import requests
 
-from temp import (
-    extracted_directory,
-    output_file,
-    check_source_measurements,
-    background_dir,
-)
-
-from libra_toolbox.neutron_detection.activation_foils.compass import (
-    Measurement,
-    CheckSourceMeasurement,
-    SampleMeasurement,
-)
-
 
 def download_and_extract_foil_data(extracted_directory: Path, output_file: Path):
 
@@ -49,31 +36,3 @@ def download_and_extract_foil_data(extracted_directory: Path, output_file: Path)
         # Delete the zip file after extraction
         output_file.unlink(missing_ok=True)
 
-
-def read_measurement_from_directory(
-    check_source_measurements: dict, background_dir: Path
-):
-
-    all_measurements = {}
-
-    for name, values in check_source_measurements.items():
-        print(f"Processing {name}...")
-        meas = CheckSourceMeasurement.from_directory(values["directory"], name=name)
-        meas.check_source = values["check_source"]
-        all_measurements[name] = meas
-
-    print(f"Processing background...")
-    background_meas = Measurement.from_directory(
-        background_dir,
-        name="Background",
-        info_file_optional=True,
-    )
-    return all_measurements, background_meas
-
-
-download_and_extract_foil_data(extracted_directory, output_file)
-
-
-check_source_measurements, background_meas = read_measurement_from_directory(
-    check_source_measurements, background_dir
-)
